+++ src/components/Collections/CollectionCard.tsx
@@ -3,15 +3,15 @@ import { GalleryContext } from 'pages/gallery';
 import { useState, useContext, useEffect } from 'react';
 import downloadManager from 'services/downloadManager';
 import { EnteFile } from 'types/file';
-import { CollectionTile } from './styledComponents';
+import { CollectionTile, LargerCollectionTile } from './styledComponents';
 
 export default function CollectionCard(props: {
     children?: any;
     latestFile: EnteFile;
     onClick: () => void;
-    customCollectionTile?: any;
+    large?: boolean;
 }) {
-    const { latestFile: file, onClick, children, customCollectionTile } = props;
+    const { latestFile: file, onClick, children, large } = props;
 
     const [coverImageURL, setCoverImageURL] = useState(null);
     const galleryContext = useContext(GalleryContext);
@@ -28,7 +28,7 @@ export default function CollectionCard(props: {
         };
         main();
     }, [file]);
-    const UsedCollectionTile = customCollectionTile ?? CollectionTile;
+    const UsedCollectionTile = large ? LargerCollectionTile : CollectionTile;
     return (
         <UsedCollectionTile coverImgURL={coverImageURL} onClick={onClick}>
             {children}
