+++ src/hooks/useComponentScroll.tsx
@@ -1,4 +1,4 @@
-import { SCROLL_DIRECTION } from 'components/Collections/NavigationButton';
+import { SCROLL_DIRECTION } from 'components/Collections/CollectionBar/NavigationButton';
 import { useRef, useState, useEffect } from 'react';
 
 export default function useComponentScroll({
